feat: move condition props to BpmnPropertiesProvider
Closes #52
