+++ src/provider/bpmn/properties/index.js
@@ -5,3 +5,4 @@ export { DocumentationProps } from './DocumentationProps';
 export { ProcessProps } from './ProcessProps';
 export { ErrorProps } from './ErrorProps';
 export { MessageProps } from './MessageProps';
+export { ConditionProps } from './ConditionProps';
