+++ src/provider/zeebe/properties/index.js
@@ -1,6 +1,6 @@
 export { InputProps } from './InputProps';
 export { OutputProps } from './OutputProps';
-export { ConditionProps } from './ConditionProps';
+export { ConditionProps } from '../../bpmn/properties/ConditionProps';
 export { MultiInstanceProps } from './MultiInstanceProps';
 export { OutputPropagationProps } from './OutputPropagationProps';
 export { TargetProps } from './TargetProps';
