+++ test/spec/provider/bpmn/BpmnPropertiesProvider.spec.js
@@ -60,6 +60,40 @@ describe('<BpmnPropertiesProvider>', function() {
   }));
 
 
+  it('should NOT show condition group', inject(async function(elementRegistry, selection) {
+
+    // given
+    const task = elementRegistry.get('Flow_1');
+
+    await act(() => {
+      selection.select(task);
+    });
+
+    // when
+    const conditionGroup = getGroup(container, 'condition');
+
+    // then
+    expect(conditionGroup).to.not.exist;
+  }));
+
+
+  it('should show condition group', inject(async function(elementRegistry, selection) {
+
+    // given
+    const serviceTask = elementRegistry.get('Flow_2');
+
+    await act(() => {
+      selection.select(serviceTask);
+    });
+
+    // when
+    const conditionGroup = getGroup(container, 'condition');
+
+    // then
+    expect(conditionGroup).to.exist;
+  }));
+
+
   it('should NOT show error group', inject(async function(elementRegistry, selection) {
 
     // given
