+++ test/spec/provider/bpmn/ConditionProps.bpmn
