+++ test/spec/provider/zeebe/ZeebePropertiesProvider.spec.js
@@ -256,40 +256,6 @@ describe('<ZeebePropertiesProvider>', function() {
   }));
 
 
-  it('should NOT show condition group', inject(async function(elementRegistry, selection) {
-
-    // given
-    const task = elementRegistry.get('Flow_1');
-
-    await act(() => {
-      selection.select(task);
-    });
-
-    // when
-    const conditionGroup = getGroup(container, 'condition');
-
-    // then
-    expect(conditionGroup).to.not.exist;
-  }));
-
-
-  it('should show condition group', inject(async function(elementRegistry, selection) {
-
-    // given
-    const serviceTask = elementRegistry.get('Flow_2');
-
-    await act(() => {
-      selection.select(serviceTask);
-    });
-
-    // when
-    const conditionGroup = getGroup(container, 'condition');
-
-    // then
-    expect(conditionGroup).to.exist;
-  }));
-
-
   it('should NOT show output propagation group', inject(async function(elementRegistry, selection) {
 
     // given
